Initial DataCite registrar implementation
Delete metadata client method added for transitioning findable to
registered.

Writing this registrar leads me to think that we should actually use the
DataCite REST API instead of the MDS API because the MDS API doesn't
allow for directly setting the state.  This leads to the odd behavior of
the registrar where delete_metadata has to be called after the URL is registered
in order to avoid the DOI being left findable when it should be registered.
Additionally the REST API may allow us to make one API call instead of
up to 4 which would be much more efficient.

It may be good to add integration tests in the future which actually
call out to a running instance of the Datacite APIs in docker containers
in CircleCI.